lab1 + bonus

+++ src/main/java/mk/finki/ukim/mk/lab/model/EventBooking.java
@@ -11,5 +11,5 @@ public class EventBooking {
     String eventName;
     String attendeeName;
     String attendeeAddress;
-    Long numberOfTickets;
+    int numberOfTickets;
 }
